% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftchart.R
\name{lift_chart_data}
\alias{lift_chart_data}
\title{Calculate Lift Chart Data}
\usage{
lift_chart_data(data, actual, predicted, groups = 10)
}
\arguments{
\item{data}{A data frame containing the actual and predicted values.}

\item{actual}{The name of the actual (dependent variable) column as a string.}

\item{predicted}{The name of the predicted scores/probabilities column as a string.}

\item{groups}{Number of groups (deciles). Default is 10.}
}
\value{
A data frame with lift chart calculations.
}
\description{
Calculate Lift Chart Data
}
\examples{
data <- data.frame(actual = rbinom(100, 1, 0.6), predicted = runif(100))
lift_data <- lift_chart_data(data, actual = "actual", predicted = "predicted")
print(lift_data)
}
