% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftchart.R
\name{plot_cumulative_gain}
\alias{plot_cumulative_gain}
\title{Plot Cumulative Gains Chart}
\usage{
plot_cumulative_gain(gain_table)
}
\arguments{
\item{gain_table}{Data frame returned by \code{lift_chart_data}.}
}
\value{
A ggplot object.
}
\description{
Plot Cumulative Gains Chart
}
\examples{
data <- data.frame(actual = rbinom(100, 1, 0.6), predicted = runif(100))
lift_data <- lift_chart_data(data, actual = "actual", predicted = "predicted")
plot_cumulative_gain(lift_data)
}
